refactor: extract function selector resolvers to resolve.rs
